BUG/MINOR: stats-file: Use a 16bits variable when loading tgid
Now that the tgid stored in the stats file has been increased to 16bits
by commit 022cb3ab7fdce74de2cf24bea865ecf7015e5754, don't forget to
increase the variable size when reading it from the file, too.
This should have no impact given the maximum thread group limit is still
32.

+++ src/stats-file.c
@@ -739,7 +739,7 @@ static void shm_stats_file_preload(void)
 	while (it < objects) {
 		struct guid_node *node;
 		uint64_t users;
-		uint8_t obj_tgid;
+		uint16_t obj_tgid;
 
 		curr_obj = SHM_STATS_FILE_OBJECT(shm_stats_file_hdr, it);
